build: generate the EPK bundle at build time instead of committing it
Split scripts/generate-epk.mjs into generate-epk-photos.mjs (resizes the
gitignored masters — local-only, output committed) and
generate-epk-bundle.mjs (PDF + CREDITS + ZIP from committed data and
photos — no masters, so CI can run it), sharing epk-context.mjs.

The CI Build job now regenerates the bundle before building, so the
downloadable PDF/ZIP are always fresh from the data (no stale one-sheet,
no manual epk:kit + commit) and the ~6 MB zip blob leaves git history.
Only the hi-res photos stay committed; E2E and the Pages deploy reuse
that same dist, so what ships is what was tested.
