+++ .gitignore
@@ -53,6 +53,11 @@ lighthouse-report*
 # Build-time responsive image variants (generated by scripts/generate-responsive-images.mjs)
 public/images/responsive/
 
+# EPK bundle — regenerated at build from committed data + photos (the photos themselves stay committed)
+public/epk/*.pdf
+public/epk/*.zip
+public/epk/CREDITS.txt
+
 # Local, private notes — planning docs kept alongside the code but never published
 .notes/
