--- public/epk/CREDITS.txt
@@ -1,11 +1,0 @@
-Jerome Faria — press kit
-
-Photography:
-  jerome-faria-1-by-pedro-jafuno.jpg — Pedro Jafuno
-  jerome-faria-2-by-pedro-jafuno.jpg — Pedro Jafuno
-  jerome-faria-3-by-pedro-jafuno.jpg — Pedro Jafuno
-  jerome-faria-4-by-pedro-jafuno.jpg — Pedro Jafuno
-  jerome-faria-5-by-ricardo-almeida.jpg — Ricardo Almeida
-  jerome-faria-6-by-pedro-roque.jpg — Pedro Roque
-
-Site content © Jerome Faria. Photographs remain the property of their authors.
